完成

+++ Mall/src/main/java/com/example/demo/model/CmsPrefrenceAreaProductRelation.java
@@ -2,7 +2,7 @@ package com.example.demo.model;
 
 //优选专区和商品的关系
 public class CmsPrefrenceAreaProductRelation {
-	private Integer id;
-	private Integer prefrenceAreaId;
-	private Integer productId;
+	private Long id;
+	private Long prefrenceAreaId;
+	private Long productId;
 }
