+++ Mall/src/main/java/com/example/demo/model/PmsProductAttributeValue.java
@@ -1,8 +1,8 @@
 package com.example.demo.model;
 
 public class PmsProductAttributeValue {
-	private Integer id;
-	private Integer productAttributeId;
-	private Integer productId;
+	private Long id;
+	private Long productAttributeId;
+	private Long productId;
 	private String value;
 } // 手动添加规格或参数的值,参数单值，规格有多个时以逗号隔开
