+++ Mall/src/main/java/com/example/demo/model/PmsProductLadder.java
@@ -3,9 +3,9 @@ package com.example.demo.model;
 public class PmsProductLadder {
 	private Double discount;//折扣
 	private Integer count;//满足的商品数量
-	private Integer id;
+	private Long id;
 	private Double price;//折后价格
-	private Integer productId;
+	private Long productId;
 
 	
 }
